FIXED: Compiler warnings, patch by Mark Rogers
git-svn-id: https://svn.code.sf.net/p/podofo/code/podofo/trunk@1740 e6e4d1b7-ffb3-4f57-9d89-6d52b3892af5

+++ src/doc/PdfFontTTFSubset.cpp
@@ -491,7 +491,7 @@ void PdfFontTTFSubset::CreateCmapTable( const CodePointToGid& usedCodes )
         //std::cout << " sc=" << std::hex << range.startCode << " ec=" << range.endCode << " dc=" << range.startCode + range.delta << " of=" << range.offset << std::dec << std::endl;
         m_sCMap.ranges.push_back(range);
                 }
-    m_sCMap.segCount = m_sCMap.ranges.size() + 1;
+    m_sCMap.segCount = static_cast<unsigned short>(m_sCMap.ranges.size() + 1);
     /* fill glyphArray */
     if (arrayCount) {
         m_sCMap.glyphArray.reserve(arrayCount);
